refactor(rc): move packages for cmd -> comp -> pkg
Imports should be cmd -> comp -> pkg, and the service init is in pkg,
meaning any code that it references should be in pkg too.
